few corrections

+++ src/vec3.cpp
@@ -1,20 +1,17 @@
 #include "vec3.h"
 #include <cassert>
 
-PhysiK::vec3::vec3(const unsigned int fst, const unsigned int snd, const unsigned int thr):
-	positions{fst,snd,thr}
-{
+PhysiK::vec3::vec3(const float fst, const float snd, const float thr):
+	positions{fst,snd,thr}{
 
 }
 
-const unsigned int& PhysiK::vec3::operator [](const unsigned int offset) const
-{
+const float& PhysiK::vec3::operator [](const unsigned int offset) const{
 	assert(offset<3);
 	return positions[offset];
 }
 
-unsigned int& PhysiK::vec3::operator [](const unsigned int offset)
-{
+float& PhysiK::vec3::operator [](const unsigned int offset){
 	assert(offset<3);
 	return positions[offset];
 }
